requirements.txt
aiofiles==23.2.1
aiohttp==3.9.2
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
bitarray==2.9.2
bitstring==4.1.4
blinker==1.7.0
blspy==2.0.2
boto3==1.34.46
botocore==1.34.108
cached-property==1.5.2
cffi==1.16.0
charset-normalizer==3.3.2
chia-blockchain==2.3.0
chia_dev_tools==1.2.6
chia_rs==0.6.1
chiabip158==1.4
chiapos==2.0.4
chiavdf==1.1.4
click==8.1.3
clvm==0.9.9
clvm_tools==0.4.9
clvm_tools_rs==0.1.40
colorama==0.4.6
colorlog==6.8.2
concurrent-log-handler==0.9.25
cryptography==42.0.5
cytoolz==0.12.3
dnslib==0.9.24
dnspython==2.5.0
eth-hash==0.7.0
eth-typing==4.2.1
eth-utils==4.1.0
exceptiongroup==1.2.0
fastapi==0.110.2
filelock==3.13.1
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
idna==3.6
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.3
jmespath==1.0.1
keyring==24.3.1
MarkupSafe==2.1.5
more-itertools==10.2.0
multidict==6.0.5
npm==0.1.1
optional-django==0.1.0
packaging==23.2
pluggy==1.4.0
portalocker==2.8.2
psutil==5.9.4
py-ecc==7.0.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
pytimeparse==1.1.8
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
s3transfer==0.10.1
setproctitle==1.3.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
tomli==2.0.1
toolz==0.12.1
typing_extensions==4.10.0
urllib3==2.0.7
uvicorn==0.29.0
watchdog==4.0.0
Werkzeug==3.0.2
yarl==1.9.4
zipp==3.18.1
zstd==1.5.5.1
